backend/requirements.txt
# FastAPI e servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Banco de dados
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # Para PostgreSQL (opcional)

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
PyJWT==2.8.0

# Utilitários
pydantic==2.5.3
python-dotenv==1.0.0